+++ src/frontend/src/App.js
@@ -9,7 +9,7 @@ import axios from 'axios'
 import Carousel from 'react-multi-carousel'
 import 'react-multi-carousel/lib/styles.css';
 
-import Movie from './Movie'
+import { Movie } from './Movie'
 import Ratings from './Ratings'
 import Wishlist from './WishList'
