+++ src/frontend/src/Cookies.js
@@ -1,36 +1,45 @@
 
 
-const setCookie = ({movieid, rating, exdays}) => {
+//Creates a new cookie with the movieid as the name of the cookie, 
+//rating as the value of the cookie. Exdays is the amount of days until the cookie expires
+function setCookie(movieid, rating, exdays) {
   const d = new Date();
   d.setTime(d.getTime() + (exdays * 24 * 60 * 60 * 1000));
-  let expires = "expires="+d.toUTCString();
-  document.cookie = movieid + "=" + rating + ";" + expires + ";path=/";
-}
-  
-const getCookie = ({movieid}) => {
-    let name = movieid + "=";
-    let ca = document.cookie.split(';');
-    for(let i = 0; i < ca.length; i++) {
-      let c = ca[i];
-      while (c.charAt(0) == ' ') {
-        c = c.substring(1);
-      }
-      if (c.indexOf(name) == 0) {
-        return c.substring(name.length, c.length);
-      }
-    }
-    return "";
+  let expires = "expires="+d.toUTCString();
+  document.cookie = "m" + movieid + "=" + rating + ";" + expires + ";path=/";
+  window.location.reload()
 }
 
-const getCookies = () =>{
+//Searchers saved cookies for a cookie with the name movieid
+//Returns rating associated with that cookie or 0 if no cookie is found
+function getCookie(movieid) {
+  let pairs = document.cookie.split(';');
+  for(let i = 0; i < pairs.length; i++) {
+    let pair = pairs[i].split('=');
+    if(pair[0].substring(0, 1) === ' '){
+      pair[0] = pair[0].substring(1)
+    }
+    if(pair[0].substring(1) === movieid){
+      return pair[1];
+    }
+  }
+  return 0;
+}
+
+//Returns a list of pairs. A pair has a cookies name/movieid in pair[0]
+//and value/rating in pair[1]
+var getCookies = function(){
   var pairs = document.cookie.split(";")
   var cookies = []
-  console.log('hei')
   for(var i=0; i<pairs.length; i++){
     var pair = pairs[i].split("=")
+    if(pair[0].substring(0, 1) === ' '){
+      pair[0] = pair[0].substring(1)
+    }
+    pair[0] = pair[0].substring(1)
     cookies[i] = pair;
   }
-    return cookies;
+  return cookies;
 }
 
-  export default { setCookie, getCookie, getCookies };
+  export { setCookie, getCookie, getCookies };
