+++ src/frontend/src/Movie.js
@@ -2,9 +2,9 @@ import axios from 'axios'
 import { useState, useEffect } from 'react'
 import ReactStars from "react-rating-stars-component"
 import  image from "./NoImage.jpg";
+import { getCookie, setCookie } from './Cookies.js'
 
 const GetMovieByID = (id) => {
-
   const [movie, setMovie] = useState([])
   useEffect(() => {   axios
     .get(`http://128.214.253.51:3000/dbgetgivenmoviedata?movieid=${id}`)
@@ -14,33 +14,6 @@ const GetMovieByID = (id) => {
   }, []);
   return (movie)
 }
-
-//Creates a new cookie with the movieid as the name of the cookie, 
-//rating as the value of the cookie. Exdays is the amount of days until the cookie expires
-function setCookie(movieid, rating, exdays) {
-  const d = new Date();
-  d.setTime(d.getTime() + (exdays * 24 * 60 * 60 * 1000));
-  let expires = "expires="+d.toUTCString();
-  document.cookie = "m" + movieid + "=" + rating + ";" + expires + ";path=/";
-  window.location.reload()
-}
-
-//Searchers saved cookies for a cookie with the name movieid
-//Returns rating associated with that cookie or 0 if no cookie is found
-function getCookie(movieid) {
-  let pairs = document.cookie.split(';');
-  for(let i = 0; i < pairs.length; i++) {
-    let pair = pairs[i].split('=');
-    if(pair[0].substring(0, 1) === ' '){
-      pair[0] = pair[0].substring(1)
-    }
-    if(pair[0].substring(1) === movieid){
-      return pair[1];
-    }
-  }
-  return 0;
-}
-
 
 const Movie = () => {
   //Gets the movieid from the url
@@ -99,4 +72,4 @@ const Movie = () => {
   )
 }
 
-export default Movie
+export { Movie, GetMovieByID }
